Split session view streaming merge helper
